Replaced class-based interface with functional interface for cl-imgmagick
cl-imgmagick originally had the image class that wrapped the MagickWand wand
C data structure. The garbage collector did not work aggressively enough for this
approach to work, since most of the memory was allocated and held by the C side
(MagickWand). Therefore, this class-based interface is being scrapped in favour
of a functional interface where the MagickWand wand data structure is only kept
around for as long as it is needed, without relying on garbage collection.

- FFI methods were added for dealing with MagickWand image list iterators.
- Either libWand.dylib or libMagickWand.dylib should be loaded on OS X. By
  default, only libWand.dylib is accessible. libMagickWand.dylib is only
  accessible if installed using MacPorts and the LD_LIBRARY_PATH equivalent
  contains /opt/local/lib.

- The image class was removed and its functionality was replaced by
  with-magick-wand and with-image-pixels macros. The methods previously defined
  for the image class were converted into functions that operate on the
  MagickWand wand.

- imgdb-store was modified to use the cl-imgmagick functional interface.

- cl-imgmagick's export list was updated accordingly.