(index-img-drop) handles indexing of image store
(index-img-drop) can now handle the indexing of the image store itself, an
image drop with duplicates or an image drop with images that already exist in
the image store. This was done by checking the digest of the file being
indexing with similarly named and dated files already in the store.